# discovery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User-facing read-only API for search and content discovery. Returns a unified list of programs and episodes. Program and episode detail endpoints return full content with optional episode pagination. Responses are cached with a short TTL for scale.",
        "title": "Discovery API",
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/search": {
            "get": {
                "description": "Retrieves a unified, relevance-ranked list of programs and episodes matching the query. Each item includes a type (\"program\" or \"episode\") and view fields only — no videoUrl is exposed in search results. Optional filters narrow by category or language. Results are paginated; use page and size to navigate. Responses are cached for 2 minutes for scale.",
                "produces": ["application/json"],
                "tags": ["search"],
                "summary": "Search programs and episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full-text search query across titles and descriptions",
                        "name": "q",
                        "in": "query",
                        "x-example": "فنجان"
                    },
                    {
                        "type": "string",
                        "description": "Filter results by category (exact match)",
                        "name": "category",
                        "in": "query",
                        "x-example": "Culture"
                    },
                    {
                        "type": "string",
                        "description": "Filter results by language code (e.g. ar, en)",
                        "name": "language",
                        "in": "query",
                        "x-example": "ar"
                    },
                    {
                        "type": "integer",
                        "description": "Page number, starts at 1",
                        "name": "page",
                        "in": "query",
                        "default": 1,
                        "x-example": 1
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (1–100, default 10)",
                        "name": "size",
                        "in": "query",
                        "default": 10,
                        "x-example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully. Items are a flat mixed list of programs and episodes.",
                        "schema": { "$ref": "#/definitions/model.SearchResult" },
                        "examples": {
                            "application/json": {
                                "items": [
                                    {
                                        "type": "program",
                                        "id": "123e4567-e89b-12d3-a456-426614174000",
                                        "title": "فنجان",
                                        "description": "A cultural podcast exploring ideas and conversations",
                                        "programType": "podcast",
                                        "category": "Culture",
                                        "language": "ar",
                                        "coverImageUrl": "https://cdn.example.com/covers/123e4567-cover.jpg"
                                    },
                                    {
                                        "type": "episode",
                                        "id": "223e4567-e89b-12d3-a456-426614174000",
                                        "title": "الحلقة الأولى: البداية",
                                        "description": "The first episode of the series",
                                        "programId": "123e4567-e89b-12d3-a456-426614174000",
                                        "thumbnailUrl": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg",
                                        "durationInSeconds": 3600,
                                        "episodeNumber": 1,
                                        "seasonNumber": 1,
                                        "publicationDate": "2024-02-17T14:30:00.000Z"
                                    }
                                ],
                                "total": 42,
                                "page": 1,
                                "limit": 10
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error — Elasticsearch or cache unavailable",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "search failed" }
                        }
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "description": "Retrieves a single program by UUID with a paginated list of its episodes ordered by season and episode number. Use episodePage and episodeSize to page through episodes (e.g. load 20 at a time on the program detail page). Internal fields (status, createdAt, updatedAt) are omitted — this is the public-facing view. Responses are cached for 3 minutes.",
                "produces": ["application/json"],
                "tags": ["programs"],
                "summary": "Get program by ID with episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Program UUID (e.g. 123e4567-e89b-12d3-a456-426614174000)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "x-example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    {
                        "type": "integer",
                        "description": "Episode list page number (starts at 1)",
                        "name": "episodePage",
                        "in": "query",
                        "default": 1,
                        "x-example": 1
                    },
                    {
                        "type": "integer",
                        "description": "Number of episodes per page (default 20, max 100)",
                        "name": "episodeSize",
                        "in": "query",
                        "default": 20,
                        "x-example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program retrieved successfully with its episode list. Episodes are ordered by seasonNumber then episodeNumber ascending.",
                        "schema": { "$ref": "#/definitions/model.ProgramDetail" },
                        "examples": {
                            "application/json": {
                                "id": "123e4567-e89b-12d3-a456-426614174000",
                                "title": "فنجان",
                                "description": "A cultural podcast exploring ideas and conversations",
                                "type": "podcast",
                                "category": "Culture",
                                "language": "ar",
                                "coverImageUrl": "https://cdn.example.com/covers/123e4567-cover.jpg",
                                "episodes": [
                                    {
                                        "id": "223e4567-e89b-12d3-a456-426614174000",
                                        "title": "الحلقة الأولى: البداية",
                                        "thumbnailUrl": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg",
                                        "videoUrl": "https://cdn.example.com/episodes/223e4567/episode.mp4",
                                        "durationInSeconds": 3600,
                                        "episodeNumber": 1,
                                        "seasonNumber": 1,
                                        "publicationDate": "2024-02-17T14:30:00.000Z"
                                    },
                                    {
                                        "id": "323e4567-e89b-12d3-a456-426614174000",
                                        "title": "الحلقة الثانية: التعمق",
                                        "thumbnailUrl": "https://cdn.example.com/thumbnails/323e4567-thumb.jpg",
                                        "videoUrl": "https://cdn.example.com/episodes/323e4567/episode.mp4",
                                        "durationInSeconds": 2700,
                                        "episodeNumber": 2,
                                        "seasonNumber": 1,
                                        "publicationDate": "2024-02-24T14:30:00.000Z"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request — id path parameter is missing or empty",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "id required" }
                        }
                    },
                    "404": {
                        "description": "Program not found — no program exists with the given UUID",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "program not found" }
                        }
                    },
                    "500": {
                        "description": "Internal server error — Elasticsearch or cache unavailable",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "internal error" }
                        }
                    }
                }
            }
        },
        "/episodes/{id}": {
            "get": {
                "description": "Retrieves a single episode by UUID including full detail (video URL, duration, season/episode numbers, publication date) and a brief summary of the parent program (id, title, cover image). Use this on the episode player page. Internal fields are omitted. Responses are cached for 3 minutes.",
                "produces": ["application/json"],
                "tags": ["episodes"],
                "summary": "Get episode by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Episode UUID (e.g. 223e4567-e89b-12d3-a456-426614174000)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "x-example": "223e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode retrieved successfully. The program field contains a lightweight program summary suitable for rendering a back-link or breadcrumb.",
                        "schema": { "$ref": "#/definitions/model.EpisodeDetail" },
                        "examples": {
                            "application/json": {
                                "id": "223e4567-e89b-12d3-a456-426614174000",
                                "title": "الحلقة الأولى: البداية",
                                "description": "The first episode of the series — an in-depth introduction",
                                "videoUrl": "https://cdn.example.com/episodes/223e4567/episode.mp4",
                                "thumbnailUrl": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg",
                                "durationInSeconds": 3600,
                                "episodeNumber": 1,
                                "seasonNumber": 1,
                                "publicationDate": "2024-02-17T14:30:00.000Z",
                                "program": {
                                    "id": "123e4567-e89b-12d3-a456-426614174000",
                                    "title": "فنجان",
                                    "coverImageUrl": "https://cdn.example.com/covers/123e4567-cover.jpg"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request — id path parameter is missing or empty",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "id required" }
                        }
                    },
                    "404": {
                        "description": "Episode not found — no episode exists with the given UUID",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "episode not found" }
                        }
                    },
                    "500": {
                        "description": "Internal server error — Elasticsearch or cache unavailable",
                        "schema": { "$ref": "#/definitions/object" },
                        "examples": {
                            "application/json": { "error": "internal error" }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "object": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "description": "Human-readable error message"
                }
            }
        },
        "model.SearchItem": {
            "type": "object",
            "description": "A single item in unified search results. Always has type, id, title. Program-only fields (programType, category, language, coverImageUrl) are present when type=program. Episode-only fields (programId, thumbnailUrl, durationInSeconds, episodeNumber, seasonNumber, publicationDate) are present when type=episode.",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Discriminator: \"program\" or \"episode\"",
                    "enum": ["program", "episode"],
                    "example": "program"
                },
                "id": {
                    "type": "string",
                    "description": "UUID of the program or episode",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                    "type": "string",
                    "description": "Title of the program or episode",
                    "example": "فنجان"
                },
                "description": { "type": "string", "description": "Short description (omitted if not set)" },
                "programType": {
                    "type": "string",
                    "description": "Program type — only present when type=program",
                    "enum": ["podcast", "documentary", "series"]
                },
                "category": {
                    "type": "string",
                    "description": "Program category — only present when type=program",
                    "example": "Culture"
                },
                "language": {
                    "type": "string",
                    "description": "Language code — only present when type=program",
                    "example": "ar"
                },
                "coverImageUrl": {
                    "type": "string",
                    "description": "Program cover image URL — only present when type=program",
                    "example": "https://cdn.example.com/covers/123e4567-cover.jpg"
                },
                "programId": {
                    "type": "string",
                    "description": "Parent program UUID — only present when type=episode",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "thumbnailUrl": {
                    "type": "string",
                    "description": "Episode thumbnail URL — only present when type=episode",
                    "example": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg"
                },
                "durationInSeconds": {
                    "type": "integer",
                    "description": "Episode duration in seconds — only present when type=episode",
                    "example": 3600
                },
                "episodeNumber": {
                    "type": "integer",
                    "description": "Episode number within its season — only present when type=episode",
                    "example": 1
                },
                "seasonNumber": {
                    "type": "integer",
                    "description": "Season number — only present when type=episode (null for single-season programs)",
                    "example": 1
                },
                "publicationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "ISO 8601 publication date — only present when type=episode",
                    "example": "2024-02-17T14:30:00.000Z"
                }
            }
        },
        "model.SearchResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "description": "Mixed list of programs and episodes ordered by relevance",
                    "items": { "$ref": "#/definitions/model.SearchItem" }
                },
                "total": {
                    "type": "integer",
                    "description": "Total number of matching results across all pages",
                    "example": 42
                },
                "page": {
                    "type": "integer",
                    "description": "Current page number",
                    "example": 1
                },
                "limit": {
                    "type": "integer",
                    "description": "Number of items returned in this page",
                    "example": 10
                }
            }
        },
        "model.EpisodeSummary": {
            "type": "object",
            "description": "Lightweight episode representation used in the program detail episode list",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Episode UUID",
                    "example": "223e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                    "type": "string",
                    "description": "Episode title",
                    "example": "الحلقة الأولى: البداية"
                },
                "thumbnailUrl": {
                    "type": "string",
                    "description": "Episode thumbnail URL",
                    "example": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg"
                },
                "videoUrl": {
                    "type": "string",
                    "description": "Streamable video URL (CDN)",
                    "example": "https://cdn.example.com/episodes/223e4567/episode.mp4"
                },
                "durationInSeconds": {
                    "type": "integer",
                    "description": "Episode duration in seconds",
                    "example": 3600
                },
                "episodeNumber": {
                    "type": "integer",
                    "description": "Episode number within its season",
                    "example": 1
                },
                "seasonNumber": {
                    "type": "integer",
                    "description": "Season number (null for single-season programs)",
                    "example": 1
                },
                "publicationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "ISO 8601 publication date",
                    "example": "2024-02-17T14:30:00.000Z"
                }
            }
        },
        "model.ProgramDetail": {
            "type": "object",
            "description": "Full program detail with paginated episode list. Internal fields (status, createdAt, updatedAt) are omitted.",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Program UUID",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                    "type": "string",
                    "description": "Program title",
                    "example": "فنجان"
                },
                "description": {
                    "type": "string",
                    "description": "Program description (omitted if not set)"
                },
                "type": {
                    "type": "string",
                    "description": "Program type",
                    "enum": ["podcast", "documentary", "series"],
                    "example": "podcast"
                },
                "category": {
                    "type": "string",
                    "description": "Program category",
                    "example": "Culture"
                },
                "language": {
                    "type": "string",
                    "description": "Language code",
                    "example": "ar"
                },
                "coverImageUrl": {
                    "type": "string",
                    "description": "Cover image URL (CDN)",
                    "example": "https://cdn.example.com/covers/123e4567-cover.jpg"
                },
                "episodes": {
                    "type": "array",
                    "description": "Paginated episode list ordered by seasonNumber then episodeNumber ascending",
                    "items": { "$ref": "#/definitions/model.EpisodeSummary" }
                }
            }
        },
        "model.ProgramSummary": {
            "type": "object",
            "description": "Lightweight program representation embedded inside EpisodeDetail for breadcrumb/back-link rendering",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Program UUID",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                    "type": "string",
                    "description": "Program title",
                    "example": "فنجان"
                },
                "coverImageUrl": {
                    "type": "string",
                    "description": "Program cover image URL (CDN)",
                    "example": "https://cdn.example.com/covers/123e4567-cover.jpg"
                }
            }
        },
        "model.EpisodeDetail": {
            "type": "object",
            "description": "Full episode detail including video URL and parent program summary. Internal fields (status, createdAt, updatedAt) are omitted.",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Episode UUID",
                    "example": "223e4567-e89b-12d3-a456-426614174000"
                },
                "title": {
                    "type": "string",
                    "description": "Episode title",
                    "example": "الحلقة الأولى: البداية"
                },
                "description": {
                    "type": "string",
                    "description": "Episode description (omitted if not set)"
                },
                "videoUrl": {
                    "type": "string",
                    "description": "Streamable video URL (CDN)",
                    "example": "https://cdn.example.com/episodes/223e4567/episode.mp4"
                },
                "thumbnailUrl": {
                    "type": "string",
                    "description": "Episode thumbnail URL",
                    "example": "https://cdn.example.com/thumbnails/223e4567-thumb.jpg"
                },
                "durationInSeconds": {
                    "type": "integer",
                    "description": "Episode duration in seconds",
                    "example": 3600
                },
                "episodeNumber": {
                    "type": "integer",
                    "description": "Episode number within its season",
                    "example": 1
                },
                "seasonNumber": {
                    "type": "integer",
                    "description": "Season number (null for single-season programs)",
                    "example": 1
                },
                "publicationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "ISO 8601 publication date",
                    "example": "2024-02-17T14:30:00.000Z"
                },
                "program": { "$ref": "#/definitions/model.ProgramSummary" }
            }
        }
    }
}